+++ scripts/deletion_profile.py
@@ -11,7 +11,7 @@ from rseqc import SAM
 from rseqc.cli_common import run_rscript
 
 
-def main():
+def main() -> None:
     parser = argparse.ArgumentParser(description=__doc__)
     parser.add_argument("--version", action="version", version="5.0.2")
     parser.add_argument("-i", "--input", dest="input_bam", help="Input BAM file. [required]")