+++ scripts/sc_seqLogo.py
@@ -12,7 +12,7 @@ import sys
 from rseqc import fastq
 
 
-def main():
+def main() -> None:
     parser = argparse.ArgumentParser(description=__doc__)
     parser.add_argument("--version", action="version", version="5.0.2")
     parser.add_argument(