+++ scripts/split_paired_bam.py
@@ -10,7 +10,7 @@ import pysam
 from rseqc.SAM import _pysam_iter
 
 
-def main():
+def main() -> None:
     parser = argparse.ArgumentParser(description=__doc__)
     parser.add_argument("--version", action="version", version="5.0.2")
     parser.add_argument(